requirements.txt
click==8.1.6
python_magic==0.4.27
